Move EPO to its correct place
