+++ robozonky-api/src/main/java/com/github/robozonky/api/notifications/WeeklySummaryEvent.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 The RoboZonky Project
+ * Copyright 2020 The RoboZonky Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,8 +15,6 @@
  */
 
 package com.github.robozonky.api.notifications;
-
-import com.github.robozonky.api.strategies.ExtendedPortfolioOverview;
 
 /**
  * Fired immediately after RoboZonky's assembled a weekly portfolio summary.
