+++ robozonky-notifications/src/main/java/com/github/robozonky/notifications/listeners/WeeklySummaryEventListener.java
@@ -20,8 +20,8 @@ import static java.util.Map.entry;
 
 import java.util.Map;
 
+import com.github.robozonky.api.notifications.ExtendedPortfolioOverview;
 import com.github.robozonky.api.notifications.WeeklySummaryEvent;
-import com.github.robozonky.api.strategies.ExtendedPortfolioOverview;
 import com.github.robozonky.notifications.AbstractTargetHandler;
 import com.github.robozonky.notifications.SupportedListener;
