+++ robozonky-notifications/src/main/java/com/github/robozonky/notifications/samples/MyWeeklySummaryEvent.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 The RoboZonky Project
+ * Copyright 2020 The RoboZonky Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,8 +16,8 @@
 
 package com.github.robozonky.notifications.samples;
 
+import com.github.robozonky.api.notifications.ExtendedPortfolioOverview;
 import com.github.robozonky.api.notifications.WeeklySummaryEvent;
-import com.github.robozonky.api.strategies.ExtendedPortfolioOverview;
 
 public final class MyWeeklySummaryEvent extends AbstractEvent implements WeeklySummaryEvent {
