+++ robozonky-test/src/main/java/com/github/robozonky/test/AbstractRoboZonkyTest.java
@@ -34,8 +34,8 @@ import org.mockito.stubbing.Answer;
 
 import com.github.robozonky.api.Money;
 import com.github.robozonky.api.Ratio;
+import com.github.robozonky.api.notifications.ExtendedPortfolioOverview;
 import com.github.robozonky.api.remote.entities.ZonkyApiToken;
-import com.github.robozonky.api.strategies.ExtendedPortfolioOverview;
 import com.github.robozonky.internal.remote.ApiProvider;
 import com.github.robozonky.internal.remote.InvestmentResult;
 import com.github.robozonky.internal.remote.OAuth;
